==> 1. DistilBERT MODEL and SPACY.ipynb <==
from transformers import DistilBertTokenizer, DistilBertForSequenceClassification
from sklearn.metrics import accuracy_score
from sklearn.model_selection import train_test_split
import pandas as pd
import torch
import spacy

device = torch.device('cuda')

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_csv("/content/drive/MyDrive/Colab Notebooks/tweets.csv", usecols = ['text','target'])

nlp = spacy.load('en_core_web_sm')

df['text'] = df['text'].apply(lambda x: ' '.join([token.lemma_ for token in nlp(x) if not token.is_stop and token.is_alpha]))

train_texts, test_texts, train_labels, test_labels = train_test_split(df['text'], df['target'], test_size=0.2, random_state=42)

tokenizer = DistilBertTokenizer.from_pretrained('distilbert-base-uncased')
model = DistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased')

model = model.to(device)

inputs = tokenizer(train_texts.tolist(), return_tensors='pt', truncation=True, padding=True, max_length=128)
labels = torch.tensor(train_labels.tolist())

optimizer = torch.optim.AdamW(model.parameters(), lr=1e-5)

for epoch in range(1):
    for i in range(len(inputs['input_ids'])):
        input_id = inputs['input_ids'][i].to(device)
        attention_mask = inputs['attention_mask'][i].to(device)
        label = labels[i].to(device)
        optimizer.zero_grad()
        outputs = model(input_ids=input_id.unsqueeze(0), attention_mask=attention_mask.unsqueeze(0), labels=label.unsqueeze(0))
        loss = outputs.loss
        loss.backward()
        optimizer.step()

        if (i+1) % 100 == 0:
            print(f'Processed {i+1} out of {len(inputs["input_ids"])}')

test_x = tokenizer(test_texts.tolist(), return_tensors='pt', truncation=True, padding=True, max_length=128)
test_y_tensor = torch.tensor(test_labels.tolist())

model.eval()
predictions = []
for i in range(len(test_x['input_ids'])):
    input_id = test_x['input_ids'][i].to(device)
    attention_mask = test_x['attention_mask'][i].to(device)
    with torch.no_grad():
        outputs = model(input_ids=input_id.unsqueeze(0), attention_mask=attention_mask.unsqueeze(0))
    prediction = torch.argmax(outputs.logits, dim=-1)
    predictions.append(prediction.item())


accuracy = accuracy_score(test_y_tensor.tolist(), predictions)
print(f'Accuracy: {accuracy}')

test_sentence = "There is a cyclone in Florida"
test_input = tokenizer(test_sentence, return_tensors='pt', truncation=True, padding=True)
test_input = {k: v.to(device) for k, v in test_input.items()}
test_output = model(**test_input)
test_prediction = torch.argmax(test_output.logits, dim=-1)
print(f'Test sentence: "{test_sentence}" is {"a disaster" if test_prediction.item() else "not a disaster"}')

import pickle
with open('/content/drive/MyDrive/Colab Notebooks/mlmodel.pkl', 'wb') as f:
    pickle.dump(model, f)

test_sentence = "There is a cyclone in Florida"
import spacy
nlp = spacy.load('en_core_web_sm')
doc = nlp(test_sentence)
locations = [ent.text for ent in doc.ents if ent.label_ == 'GPE']
print("Disaster Locations:", locations)

test_sentence = "My life is a cyclone"
test_input = tokenizer(test_sentence, return_tensors='pt', truncation=True, padding=True)
test_input = {k: v.to(device) for k, v in test_input.items()}
test_output = model(**test_input)
test_prediction = torch.argmax(test_output.logits, dim=-1)
print(f'Test sentence: "{test_sentence}" is {"a disaster" if test_prediction.item() else "not a disaster"}')
