==> 3. LSTM MODEL & SPACY.ipynb <==
import pandas as pd
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences
from sklearn.model_selection import train_test_split
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Embedding, LSTM, Dense

from google.colab import drive
drive.mount('/content/drive')

path = "/content/drive/MyDrive/ML_Project/tweets.csv"
df = pd.read_csv(path)

df = df.dropna(subset=['text'])

tokenizer = Tokenizer()
tokenizer.fit_on_texts(df['text'])
total_words = len(tokenizer.word_index) + 1

sequences = tokenizer.texts_to_sequences(df['text'])
padded_sequences = pad_sequences(sequences)

X_train, X_test, y_train, y_test = train_test_split(padded_sequences, df['target'], test_size=0.2, random_state=42)

embedding_dim = 50
model = Sequential()
model.add(Embedding(input_dim=total_words, output_dim=embedding_dim, input_length=padded_sequences.shape[1]))
model.add(LSTM(100))
model.add(Dense(1, activation='sigmoid'))

model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])

model.fit(X_train, y_train, epochs=5, validation_data=(X_test, y_test))

loss, accuracy = model.evaluate(X_test, y_test)
print(f"Accuracy: {accuracy * 100:.2f}%")

test_sentence = ["No cows today but our local factory is sadly still ablaze"]

test_sequences = tokenizer.texts_to_sequences(test_sentence)
padded_test_sequences = pad_sequences(test_sequences, maxlen=padded_sequences.shape[1])

predictions = model.predict(padded_test_sequences)

threshold = 0.5
if predictions[0, 0] >= threshold:
    print(f'Test sentence: "{test_sentence[0]}" is a disaster.')
else:
    print(f'Test sentence: "{test_sentence[0]}" is not a disaster.')

test_sentence = "There is a cyclone in Florida"
import spacy
nlp = spacy.load('en_core_web_sm')
doc = nlp(test_sentence)
locations = [ent.text for ent in doc.ents if ent.label_ == 'GPE']
print("Disaster Locations:", locations)

import pickle
with open('/content/drive/MyDrive/ML_Project/lstm.pkl', 'wb') as f:
    pickle.dump(model, f)
