==> 4. LSTM Test.ipynb <==
import pandas as pd
import pickle
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

from google.colab import drive
drive.mount('/content/drive')

with open('/content/drive/MyDrive/ML_Project/lstm.pkl', 'rb') as f:
    model = pickle.load(f)

path = "/content/drive/MyDrive/ML_Project/tweets.csv"
df = pd.read_csv(path)

df = df.dropna(subset=['text'])

tokenizer = Tokenizer()
tokenizer.fit_on_texts(df['text'])
total_words = len(tokenizer.word_index) + 1

sequences = tokenizer.texts_to_sequences(df['text'])
padded_sequences = pad_sequences(sequences)

path = "/content/drive/MyDrive/ML_Project/fb_scraped.csv"
scraped_df = pd.read_csv(path)

scraped_df

new_tweet = "You are Amee Storm, a pink Simic Barbarian who's from the royal court and wields a blood stained tooth"
sequences = tokenizer.texts_to_sequences([new_tweet])
padded_sequences = pad_sequences(sequences, maxlen=padded_sequences.shape[1])
prediction = model.predict(padded_sequences)

threshold = 0.5

if prediction[0, 0] >= threshold:
    print(f'Tweet: "{new_tweet}" is a disaster.')
else:
    print(f'Tweet: "{new_tweet}" is not a disaster.')

for texts in scraped_df['Text']:
  new_tweet = texts
  sequences = tokenizer.texts_to_sequences([new_tweet])
  padded_sequences = pad_sequences(sequences, maxlen=padded_sequences.shape[1])
  prediction = model.predict(padded_sequences)

  threshold = 0.5

  if prediction[0, 0] >= threshold:
      print(f'Post: "{new_tweet[:35]}" is a disaster.')
  else:
      print(f'Post: "{new_tweet[:35]}" is not a disaster.')
